+++ pages/api/search/index.ts
@@ -2,13 +2,9 @@
 import axios from 'axios';
 import type { NextApiRequest, NextApiResponse } from 'next'
 
-type Data = {
-  name: string
-}
-
 export default async function handler(
   req: NextApiRequest,
-  res: NextApiResponse<Data>
+  res: NextApiResponse<{ name: string}>
 ) {
   const { q } = req.query;
