+++ types/Button/ButtonProps.ts
@@ -1,0 +1,12 @@
+export interface ButtonProps {
+  withInput?: boolean;
+  href?: string;
+  padding?: string;
+  background?: string;
+}
+
+export type ButtonElementProps = ButtonProps & {
+  color?: string;
+  children?: React.ReactNode;
+  onClick?: (e: { preventDefault: () => void; } | React.MouseEvent<HTMLButtonElement, MouseEvent>) => void;
+}
