+++ types/Card/ListingCardProps.ts
@@ -1,0 +1,11 @@
+export type ListingCardProps = {
+  id: string;
+  object: string;
+  set: string;
+  collector_number: string;
+  name: string;
+  image_uris: {
+    small: string;
+    normal: string;
+  },
+}
