+++ types/Deck/SingleDeckProps.ts
@@ -1,0 +1,5 @@
+export type SingleDeckProps = {
+  _id: string;
+  name: string;
+  deck: string;
+};
