+++ types/Session.ts
@@ -1,0 +1,6 @@
+export type SessionProps = {
+  user: {
+    email: string,
+    name: string,
+  }
+};
